requirements.txt
asgiref==3.4.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coverage==6.4.1
cryptography==38.0.2
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.4
pycparser==2.21
PyJWT==2.1.0
python-dotenv==0.21.0
pytz==2021.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
